Conflicts:
	Source/WebCore/bindings/js/JSInspectorFrontendHostCustom.cpp
	Source/WebCore/dom/WebKitNamedFlow.h
	Source/WebCore/platform/graphics/ImageSource.h

+++ Source/WebKit2/Shared/ShareableSurface.h
@@ -55,14 +55,14 @@ public:
         static bool decode(CoreIPC::ArgumentDecoder*, Handle&);
 
 #if USE(GRAPHICS_SURFACE)
-        uint32_t graphicsSurfaceToken() const { return m_graphicsSurfaceToken; }
+        uint64_t graphicsSurfaceToken() const { return m_graphicsSurfaceToken; }
 #endif
 
     private:
         friend class ShareableSurface;
         mutable ShareableBitmap::Handle m_bitmapHandle;
 #if USE(GRAPHICS_SURFACE)
-        uint32_t m_graphicsSurfaceToken;
+        uint64_t m_graphicsSurfaceToken;
 #endif
         WebCore::IntSize m_size;
         ShareableBitmap::Flags m_flags;
